refactor(web): decompose ModelDetailModal
ModelDetailModal was one 740-line component. It keeps the editor hook,
the header, the price-pin banner, the capabilities and subscription
sections, the unsaved-changes dialog and the composition (273 lines,
under the max-lines-per-function cap, so its eslint-disable goes); the
rest moves next to it under pages/Models/:

- ModelStatsGrid.tsx: the two-column facts grid with its edit inputs and
  revert buttons (RevertButton lives with it)
- ModelActionsFooter.tsx: the management footer (toggle, test, two-step
  delete, edit/save/cancel, re-discover with its cooldown)
- ModelSnippetPanel.tsx: the snippet tabs and the highlighted snippet
- modelSnippets.ts: the per-client snippet table as a pure builder
- modelDetailParse.ts: parseParams and the modalities parsing
- useModelActions.ts: the re-discovery cooldown and the test flow with
  their timers, cleared on unmount

The modal's suite passes unchanged; useModelActions gets its own
fake-timer test (30 s cooldown counting down and re-enabling, the 3 s
error flash, a thrown test, the streamed vs non-streamed success
toast). The file leaves the size allowlist (823 -> 273).

Verified on the rebuilt dev stack: a row opens the modal with its 8
facts, 10 snippet tabs (switching changes the code), edit mode shows
its 5 inputs with Save, changing a value reveals the revert buttons and
cancelling with unsaved changes opens the confirm dialog, no console
errors.
