+++ scripts/ci/size-allowlist.txt
@@ -29,6 +29,5 @@ web/src/api/types.ts	1058
 web/src/components/Layout.tsx	1065
 web/src/components/VirtualModelTable.tsx	858
 web/src/components/__tests__/Layout.navigation.test.tsx	2558
-web/src/pages/Models/ModelDetailModal.tsx	823
 web/src/pages/Settings/alerts/AlertsWizard.tsx	801
 web/src/pages/Settings/alerts/steps.tsx	846
